refactor(server): tts trait
