+++ apps/server/api/src/config/tts.rs
@@ -2,23 +2,29 @@ use serde::Deserialize;
 
 #[derive(Debug, Default, Deserialize, Clone)]
 pub struct TtsConfig {
-    model: Option<String>,
-    voice: Option<String>,
+    model: Option<Model>,
+    path: Option<String>,
 }
 
 impl TtsConfig {
     pub fn new() -> Self {
         Self {
-            model: Some(String::from("data/tts/model.onnx")),
-            voice: Some(String::from("data/tts/voice.bin")),
+            model: Some(Model::Kokoro),
+            path: Some(String::from("data/tts/mzdk100/kokoro/")),
         }
     }
 
-    pub fn model(&self) -> &str {
-        self.model.as_deref().unwrap_or_default()
+    pub fn path(&self) -> &str {
+        self.path.as_deref().unwrap_or_default()
     }
 
-    pub fn voice(&self) -> &str {
-        self.voice.as_deref().unwrap_or_default()
+    pub fn model(&self) -> Model {
+        self.model.clone().unwrap_or_default()
     }
+}
+#[derive(Clone, Debug, Deserialize, PartialEq, Default)]
+pub enum Model {
+    #[default]
+    Kokoro,
+    Voxcpm,
 }
