+++ apps/server/api/src/ws/tts/model/mod.rs
@@ -1,0 +1,2 @@
+pub mod kokoro;
+pub mod voxcpm;
