+++ apps/server/script/download_model.sh
@@ -9,6 +9,7 @@ wget https://huggingface.co/openai/whisper-large-v3-turbo/resolve/41f01f3fe87f28
 cd ..
 mkdir -p tts
 cd tts || exit
+mkdir -p tts/mzdk100/kokoro
 wget https://github.com/mzdk100/kokoro/releases/download/V1.1/kokoro-v1.1-zh.onnx -O model.onnx
 wget https://github.com/mzdk100/kokoro/releases/download/V1.1/voices-v1.1-zh.bin -O voice.bin
 cd ..
